requirements.txt
# FastAPI Backend Dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic>=2.8.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# AI and Agent Framework
smolagents==1.20.0
google-generativeai==0.3.2

# API Integrations
notion-client==2.2.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Email
# smtplib2==0.2.1  # Using built-in smtplib instead

# Scheduling and Background Tasks
celery==5.3.4
redis==5.0.1
apscheduler==3.10.4

# Utilities
python-dotenv==1.0.0
requests>=2.32.3
python-dateutil==2.8.2
pytz==2023.3

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0 

# Task Helper Tools (from colleagues)
transformers==4.53.0
sentencepiece==0.2.0
torch>=2.3.0
scipy>=1.11.0
pillow>=10.0.0
pymupdf==1.24.4
pix2tex[gui]
scikit-learn>=1.3.0

# Additional dependencies for ethics system
pandas>=2.0.0
numpy>=1.24.0 